Refactor session engine to integrate session runtime recovery and streamline event handling
